requirements.txt
langchain-community
langchain-openai
langchain
python-dotenv
flask
